+++ retail/modules/partyPointer.lua
@@ -74,8 +74,10 @@ function BBP.PartyPointer(frame)
         frame.partyPointer.healerIcon:SetVertexColor(0,1,0)
         frame.partyPointer.healerIcon:Hide()
 
+        frame.partyPointer:SetIgnoreParentAlpha(true)
         frame.partyPointer:SetFrameStrata("LOW")
     end
+    frame.partyPointer:SetAlpha(1)
 
     -- Enhanced Test Mode: Use local test variables
     if config.partyPointerTestMode then
@@ -230,6 +232,7 @@ function BBP.PartyPointer(frame)
     else
         frame.partyPointer.healerIcon:Hide()
     end
+    BBP.classIconNames[info.name] = frame
     frame.partyPointer:Show()
     if config.partyPointerHideRaidmarker then
         frame.RaidTargetFrame.RaidTargetIcon:SetAlpha(0)
@@ -251,5 +254,17 @@ function BBP.PartyPointer(frame)
         end
         BBP.hideFriendlyCastbar = nil
         frame.ppChange = nil
+    end
+end
+
+
+function BBP.FadePartyPointer(name, fade)
+    local frame = BBP.classIconNames[name]
+    if frame and frame.partyPointer then
+        if fade then
+            frame.partyPointer:SetAlpha(0.15)
+        else
+            frame.partyPointer:SetAlpha(1)
+        end
     end
 end